Bug fixes
Fixed a very weird bug, when the app would crash if the user joined a group, left the group and then joined it again.

Fixed a bug where every single inside group would update whenever someone joined or left any group

fImproved speed of someone leaving the group for the rest of the group members.